# swagger.yaml
swagger: "2.0"
info:
  title: B2c Startup backend
  description: Contains all api's and api endpoints and their descriptions.
  version: 1.0
  contact:
    email: azeemsayyad1652@gmail.com
  servers:
    description: This is a Local server
schemes: [http, https]
paths:
...
  /auth/serviceRegistration:
    post:
      summary: Service provider Registration
      parameters:
        - in: body
          name: serviceProvider
          schema:
            $ref: "#/components/schemas/serviceProviderAsParams"
      responses:
        200:
          description: User Registered Successfully
          schema:
            $ref: "#/components/schemas/serviceProviderAsResponse"
        400:
          description: Error
...
  /user/updatePicture/{user_id}:
    patch:
      summary: change Profile picture
      description: Update Profile Picture
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      - in: formData
        name: image
        type: file
      responses:
        200:
          description: Password reset successful
          schema:
            type: object
            properties:
              message:
                type: string
              updatedUser:
                $ref: "#/components/schemas/user"
...
components:
  schemas:
...
    serviceProviderAsParams:
      type: object
      properties:
        user_id:
          type: string
        profession:
          type: array
          items:
            type: string
        experience:
          type: string
        about:
          type: string
        location:
          type: string
...
    serviceProviderAsResponse:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        profession:
          type: array
          items:
            type: string
        experience:
          type: string
        profilePicture:
          type: string
        contact:
          type: string
        user_id:
          type: string
        about:
          type: string
        gallery:
          type: array
          items:
            type: string
        location:
          type: string
        notification:
          type: array
          items:
            type: string
...
    user:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        contact:
          type: string
        password:
          type: string
        profilePicture:
          type: string
...